requirements.txt
python-dateutil>=2
selenium>=3
beautifulsoup4>=4
xmljson>=0.1